refactor: move files to common package
