+++ app/src/main/java/org/feature/fox/coffee_counter/ui/EditUserView.kt
@@ -21,6 +21,7 @@ import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.launch
 import org.feature.fox.coffee_counter.R
 import org.feature.fox.coffee_counter.data.local.database.tables.User
+import org.feature.fox.coffee_counter.ui.common.CustomButton
 
 @Composable
 fun EditUserView(
