+++ app/src/main/java/org/feature/fox/coffee_counter/ui/UsersView.kt
@@ -20,6 +20,7 @@ import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.launch
 import org.feature.fox.coffee_counter.R
 import org.feature.fox.coffee_counter.data.local.database.tables.User
+import org.feature.fox.coffee_counter.ui.common.SearchBar
 
 @Preview(showSystemUi = true)
 @Composable
