+++ app/src/main/java/org/feature/fox/coffee_counter/ui/common/BottomNavigationView.kt
@@ -1,4 +1,4 @@
-package org.feature.fox.coffee_counter.ui
+package org.feature.fox.coffee_counter.ui.common
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.RowScope
@@ -21,6 +21,10 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
+import org.feature.fox.coffee_counter.ui.HistoryView
+import org.feature.fox.coffee_counter.ui.ItemsView
+import org.feature.fox.coffee_counter.ui.ProfileView
+import org.feature.fox.coffee_counter.ui.UsersView
 import org.feature.fox.coffee_counter.ui.theme.CrayolaBrown
 import org.feature.fox.coffee_counter.ui.theme.LiverOrgan
 
@@ -53,7 +57,6 @@ fun BottomNavBar(navController: NavHostController) {
         }
     }
 }
-
 
 @Composable
 fun Navigation(navController: NavHostController) {
